fix(agent): review fixes, visibility gates, and per-tool modules
Correctness:
- read_comments had no post-visibility gate; it now joins Post with the same
  predicate read_post uses, plus banned
- postVisibleInterestMatch kept its own skippedTypes and no source filter, so
  collections still became findings via the live path; both paths now share
  excludedInterestPostTypes + getExcludedInterestSourceIds, and it checks banned
- toCandidates and add_finding were missing banned/visible
- interestFindings re-validates post state, so a finding banned after delivery
  stops showing
- query_feed scope "tag" resolves synonyms; post_keyword inherits keyword.status
  via trigger, so a synonym's rows are never 'allow' and the query silently
  returned nothing
- get_source checked exclusion against the raw input, letting an excluded source
  through by handle
- write_post refuses a second call instead of orphaning a freeform post
- budget no longer gates set_interest_tags/add_finding/write_post, so a long run
  cannot end having persisted nothing
- lastRunSummary keeps the previous recap instead of being wiped on a quiet run
- offset clamp and the relevance->upvotes mapping are now reported, and
  read_comments returns real counts (postCommentCount, per-comment replyCount)
  instead of fetched ones

Structure:
- runInterestAgent.ts 1743 -> 397 lines; all 13 tools moved to
  src/common/interest/tools/, one per file, over a shared candidate pipeline
- tools/registry.ts owns membership, order and output gating; tool names derive
  from the definitions, so the list can no longer drift
- createInterestAgentTools is exported so tests can drive the real tools: pi is
  mocked under Jest, which is why none of this was reachable before

Tests: 83 passing (was 64). New coverage for the tool bodies, the registry
gating, the live-path exclusions and the read-time finding filter.
